+++ idofront-config/src/main/kotlin/com/mineinabyss/idofront/config/ConfigEntry.kt
@@ -1,0 +1,5 @@
+package com.mineinabyss.idofront.config
+
+import java.nio.file.Path
+
+data class ConfigEntry<T>(val path: Path, val data: T)