+++ idofront-config/src/main/kotlin/com/mineinabyss/idofront/config/ConfigEntryWithKey.kt
@@ -1,0 +1,9 @@
+package com.mineinabyss.idofront.config
+
+import java.nio.file.Path
+
+data class ConfigEntryWithKey<T>(
+    val path: Path,
+    val key: String,
+    val entry: T,
+)