+++ idofront-config/src/main/kotlin/com/mineinabyss/idofront/config/Format.kt
@@ -1,6 +1,0 @@
-package com.mineinabyss.idofront.config
-
-import kotlinx.serialization.StringFormat
-
-@Deprecated("Use new config api instead")
-class Format(val ext: String, val stringFormat: StringFormat)