+++ CHANGELOG.md
@@ -17,6 +17,7 @@ try to keep MAJOR changes in sync with Minecraft updates.
 ### Changed
 
 - **Feature system** has been reworked as a DSL built around Koin for dependency injection
+- **Config system** has been reworked to use a builder class that can then pick whether to decode a single file, directory, or multi-entry format which has been moved here from Geary
 - **SerializableItemStack** uses a new service for letting other plugins register custom item types instead of manually adding support for them
 - Brigadier **command api** internals reworked with much cleaner inside logic and fixes to expected behaviour from Brigadier